+++ theme.js
@@ -8,6 +8,7 @@ function loadScript(url) {
 (() => {
     loadScript("/appearance/themes/Dark+/script/module/blockattrs.js");
     loadScript("/appearance/themes/Dark+/script/module/goto.js");
+    loadScript("/appearance/themes/Dark+/script/module/reload.js");
     loadScript("/appearance/themes/Dark+/script/module/style.js");
     loadScript("/appearance/themes/Dark+/script/module/timestamp.js");